requirements.txt
dash==1.16.3
kafka-python==2.0.1
mysqlclient==2.0.1
pandas==1.1.2
plotly==4.10.0
python-dateutil==2.8.1
